# compiler: simplify calling convention