… internal/engine/compiler/compiler_conversion_test.go
… internal/engine/compiler/compiler_global_test.go
… internal/engine/compiler/compiler_initialization_test.go
… internal/engine/compiler/compiler_memory_test.go
… internal/engine/compiler/compiler_numeric_test.go
… internal/engine/compiler/compiler_post1_0_test.go
… internal/engine/compiler/compiler_stack_test.go
… internal/engine/compiler/compiler_test.go
… internal/engine/compiler/compiler_value_location_test.go
… internal/engine/compiler/compiler_vec_test.go
… internal/engine/compiler/engine_test.go
… internal/engine/compiler/impl_amd64_test.go
… internal/engine/compiler/impl_vec_amd64_test.go
… internal/engine/compiler/impl_vec_arm64_test.go
… internal/engine/interpreter/interpreter_test.go
… internal/integration_test/filecache/filecache_test.go
… internal/wasmdebug/debug_test.go
… internal/wazeroir/compiler_test.go