@mathetake 2022-08-31
ok progress 20% 😄